bot/pyproject.toml
[project]
name = "lms-telegram-bot"
version = "0.1.0"
description = "Telegram bot for LMS integration"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "python-telegram-bot>=21.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
